test invite accept update, add datum client
